requirements.txt
Django==5.0
djangorestframework
django-cors-headers
pandas
openpyxl
python-decouple
dj-database-url
psycopg2-binary
cloudinary
django-cloudinary-storage
whitenoise
gunicorn
Pillow